% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapGenerate.R
\name{heatmapGenerate}
\alias{heatmapGenerate}
\title{Generate a heatmap for gene expression in different cancer types}
\usage{
heatmapGenerate(dataset, feature, geneExpStart)
}
\arguments{
\item{feature}{the column on which to get the feature for heat map}

\item{geneExpStart}{the first column of gene expression data}

\item{data}{a dataframe containing the data for the heatmap generation}
}
\value{
NULL and prints the heatmap
}
\description{
Generate a heatmap for gene expression in different cancer types
}
\note{
ChatGPT was used to assist in searching for appropriate functions to use
}
\examples{
\dontrun{
library(TCGA)
}

}
\references{
R Core Team (2023). \emph{R: A Language and Environment for Statistical Computing}.
R Foundation for Statistical Computing, Vienna, Austria.
\url{https://www.R-project.org/}.
OpenAI. (2023). GPT-3.5: A language model by OpenAI.https://www.openai.com/gpt-3.5
}
