% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrityCheckClinical.R
\name{integrityCheckClinical}
\alias{integrityCheckClinical}
\title{Ensure that the format of the clinical dataset being passed in is of correct format}
\usage{
integrityCheckClinical(dataset)
}
\arguments{
\item{dataset}{A clinical dataset containing clinical information of interest. Must include patient IDs linking this data to the gene expression data.}
}
\value{
True if the dataset is error free. False otherwise
}
\description{
Ensure that the format of the clinical dataset being passed in is of correct format
}
\examples{
\dontrun{
library(BRCA.clinical)
integrityCheckClinical(BRCA.clinical)
}

}
\references{
R Core Team (2023). \emph{R: A Language and Environment for Statistical Computing}.
R Foundation for Statistical Computing, Vienna, Austria.
\url{https://www.R-project.org/}.
OpenAI. (2023). GPT-3.5: A language model by OpenAI.https://www.openai.com/gpt-3.5
}
