% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadataExtract.R
\name{metadataExtract}
\alias{metadataExtract}
\title{Obtain metadata about each patient through the clincal data}
\usage{
metadataExtract(dataset, metadata = colnames(dataset))
}
\arguments{
\item{dataset}{A clinical dataset containing clinical information of interest - eg. time to death, age of diagnostics. Must include patient IDs linking this data to the gene expression data.}

\item{metadata}{A vector of column names of metadata of interest. If no parameters provided, default will return the entire dataset}
}
\value{
A dataframe that contains the clinical metadata for patients
}
\description{
Obtain metadata about each patient through the clincal data
}
\examples{
\dontrun{
metadataExtract(BRCA.clinical)
}
}
