% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessRNASeq.R
\name{processRNASeq}
\alias{processRNASeq}
\title{Clean and preprocess the RNA sequencing data. Prints information about the normalisation}
\usage{
processRNASeq(dataset)
}
\arguments{
\item{dataset}{A count matrix of gene expression where each row is the expression of genes for a patient}

\item{normMethod}{Method of Normalisation}
}
\value{
A processed count matrix that was normalised.
}
\description{
Clean and preprocess the RNA sequencing data. Prints information about the normalisation
}
\examples{
\dontrun{
processRNASeq(BRCA.clinical)
}

}
