% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNASeqHeatMap.R
\name{RNASeqHeatMap}
\alias{RNASeqHeatMap}
\title{Plots a heatmap for the normalised count matrix for explanatory analysis of RNASeq expression data}
\usage{
RNASeqHeatMap(countMatrix)
}
\arguments{
\item{countMatrix}{Normalised Count Matrix}
}
\value{
returns NULL and plots heatmap for the RNA seq count matrix
}
\description{
Plots a heatmap for the normalised count matrix for explanatory analysis of RNASeq expression data
}
