% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runClinicalTranscriptLink.R
\name{runClinicalTranscriptLink}
\alias{runClinicalTranscriptLink}
\title{Launch Shiny App for ClinicalTranscriptLink}
\usage{
runClinicalTranscriptLink()
}
\value{
No return value but open up a Shiny page.
}
\description{
A function that launches the Shiny app for ClinicalTranscriptLink.
This app will allow users to create visualisations gene expression
based on different clinical attributes.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{

ClinicalTranscriptLink::runClinicalTranscriptLink()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
