% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survivalExplore.R
\name{survivalExplore}
\alias{survivalExplore}
\title{Plot histograms for distributions}
\usage{
survivalExplore(clinicalMeta, interest)
}
\arguments{
\item{clinicalMeta}{clinical metadata}

\item{interest}{factor of interest}
}
\value{
returns NULL and plots the histograms
}
\description{
Plot histograms for distributions
}
\examples{
\dontrun{
library(RTCGA.clinical)
library(BRCA.clinical)
metadata <- BRCA.clinical
metadataDeceased <- subset(metadata,
metadata$patient.clinical_cqcf.consent_or_death_status == "deceased")
metadataDeceased$patient.days_to_death <- as.numeric(metadataDeceased$patient.days_to_death)
survivalExplore(metadataDeceased, "patient.days_to_death")
}

}
\references{
R Core Team (2023). \emph{R: A Language and Environment for Statistical Computing}.
R Foundation for Statistical Computing, Vienna, Austria.
\url{https://www.R-project.org/}
}
