% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeClinicalRNA.R
\name{mergeClinicalRNA}
\alias{mergeClinicalRNA}
\title{Merge the Clinical and RNASeq data based on unique column - the patient barcode}
\usage{
mergeClinicalRNA(clinical, RNA, ID1, ID2 = ID1)
}
\arguments{
\item{clinical}{A clinical dataset containing clinical information of interest. Must include patient IDs linking this data to the gene expression data.}

\item{RNA}{A RNA-seq dataset containing gene expression}

\item{ID}{\itemize{
\item the name of the column to merge on
}}
}
\value{
The merged dataset
}
\description{
Merge the Clinical and RNASeq data based on unique column - the patient barcode
}
\note{
Generative AI - ChatGPT was used to assist in providing a base code on how
to use different functions - eg. question used was
}
\examples{
\dontrun{
TODO
}

}
\references{
R Core Team (2023). \emph{R: A Language and Environment for Statistical Computing}.
R Foundation for Statistical Computing, Vienna, Austria.
\url{https://www.R-project.org/}.
OpenAI. (2023). GPT-3.5: A language model by OpenAI.https://www.openai.com/gpt-3.5
}
